feat(@movemate/backend): renamed the entity file name
