+++ apps/backend/backend-main/src/entity/Report.ts
@@ -1,7 +1,7 @@
 import { Column, Entity, Index, ManyToOne, RelationId } from 'typeorm';
 import { BaseEntity } from './BaseEntity';
 import { User } from './User';
-import { SendPackageOrder } from './SendAPackage';
+import { SendPackageOrder } from './SendPackageOrder';
 
 @Index('IDX_report_customerId', ['customerId'], {
   where: '"deletedAt" IS NULL',
